Fixed save_or_load_embeds and moved to data loader utils
